chore: Add missing package.json for conveniet local deployment
Note that the yarn.lock is uploaded intentionally, please do not add it to .gitignore. It is generally a good practice to keep track on dependencies so they are synchronized amoung all contributors.

However I am not the maintainer of this project. If anyone would not prefer using yarn, please switch right away.